+++ src/tooling/piral-cli-xbuild/src/index.ts
@@ -1,0 +1,8 @@
+import * as actions from './actions';
+import type { CliPlugin } from 'piral-cli';
+
+const plugin: CliPlugin = (cli) => {
+  cli.withBundler('xbuild', actions);
+};
+
+module.exports = plugin;